packages/sessionize/sessionize-0.0.18-py3-none-any.whl/sessionize/utils/create.py
from typing import List, Optional

# TODO: replace with interfaces
from sqlalchemy import Table
from sqlalchemy.engine import Engine
import sqlalchemize.create as create


def create_table(
    table_name: str,
    column_names: List[str],
    column_types: List[type],
    primary_key: str,
    engine: Engine,
    schema: Optional[str] = None,
    autoincrement: Optional[bool] = True,
    if_exists: Optional[str] = 'error'
) -> Table:
    return create.create_table(table_name, column_names, column_types, primary_key, engine, schema, autoincrement, if_exists)